% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusdata.R
\docType{data}
\name{rusdata}
\alias{rusdata}
\title{Russian Subject Collection}
\format{XML}
\description{
Russian Subject Collection, Hoover Institution Library & Archives, USA
}
\references{
See \href{http://www.oac.cdlib.org/findaid/ark:/13030/tf9t1nb3q4/?query=Russian+subject+collection}{Russian Subject Collection}.
}
\keyword{Collection}
\keyword{Russian}
\keyword{Subject}
